change route name in dashboard
