--- app/(dashboard)/dashboard/order/completed-orders/page.jsx
@@ -1,12 +1,0 @@
-import CompletedOrdesPager from '@/src/dashboard/order/completedOrder';
-import React from 'react';
-
-const completedorder = () => {
-    return (
-        <div>
-           <CompletedOrdesPager/> 
-        </div>
-    );
-};
-
-export default completedorder;
