+++ app/(dashboard)/dashboard/order/processing-orders/page.jsx
@@ -1,0 +1,11 @@
+import PendingOrdersPage from "@/src/dashboard/order/pendingOrder"
+
+const pending=()=> {
+  return (
+    <div>
+        <PendingOrdersPage/>
+    </div>
+  )
+}
+
+export default pending
